requirements.txt
Flask
google-cloud-retail
google-auth
gunicorn
